Fix bug: Consecutive sections would not render
When a section is parsed and before it is refactored it is deleted from the
current set of docstring lines. However the current line pointer was not
updated. As a results after the refactoing of the section the parsing
would continue at an incorrect place. To solve the issue we set the index
at the begining of the section before refactoring and then manually
forward the pointer to the right place after the addition of the refactoed
lines