Fix qt_slider and wx_slider to properly react when the user provided an invalid value
This fix should cover the following three cases:
- The to_slider function fails
- The to_slider function does return a value that can be converted to a float
- The to_slider returns a value that is out of range

When one of these cases takes place the slider is not moved and the value
attribute is syncronized with the tollkit widget. The error and exception attributes are updated with the error information.

TODO: currently the test verifies that the last case is handled properly