Update slider interface:
- removed the converter attribute
- Slider value is now an integer
- New attributes "maximum" and "minimum" set the range of the slider

The maximum and minimum attributes are defined as Property traits that
check and validate their values so that the range is valid (i.e.
'minimum' < 'maximum'). Furthermore, the setter functions for these
properties make sure that the expected assignment and event firing
takes place when setting the slider bounds causes the current position
to change.