update the vm, parser, and compiler.
Create a new __enaml_call__ method and associated ENAML_CALL op_code to
create the components in the vm. This allows us to distinguish calling a
defn from Python (where we want to return a view) and calling a defn from
the Enaml VM (where we want to return a sequence of components and a
namespace). Created a dumb view object for use when calling defn functions
from Python.

Updated the parser and other machinery to allow a user to capture items from
the namespace of a defn call. The syntax is as follows:

    defn SomeView:
        SomeCall -> arg1, arg2, argn, some_item as foo_item, * as whole_namespace:
            foo_item.text = 'I overrode the text'

This allows a user of a predefined complex ui to reach in and change some attributes.