ENH: add :: python :: delimeter to declare a block of Python.
This commit adds better support for defining raw blocks of
standard Python code from within Enaml files. Simply bracket
the Python as follows:

:: python ::

<arbitrary python code>

:: end ::

The delimiters allow the lexer to extract python code as
a single block that can be parsed with Python's builtin compile.
This alleviates the Enaml parser from needing to understand all
of Python's syntax (and it's faster too).