BUG: fix Form layout as it was over-constraining things.
This fixes the "jumpiness" we were seeing the Form layout during
vertical resizes. The old code was setting up vertical stack
constraints on the labels and the widgets within the contrainer,
then further setting up v_center constraints for each form pair.
We only need one vertical stack on the widgets, and then v_center
for the pairs.