Update Field and tests for Field.
wx isn't quite equivalent to Qt in how it handles cursor position
and selection updates. Wx just doesn't provide enough information
and it will be necessary to write a completely custom subclass of
wx.TextCtrl to make the two entirely equivalent. The effort to do
that is just not worth it. Just use Qt.