add 'traverse' and 'find_by_name', and the 'name' attribute on BaseComponent.
This trio assists in running the unit tests for the new grammar as well
as making it easy for users to find a component by name. Note that this
is distinctly different from using an identifier, since multiple identifiers
may be in use for a given object in different scopes.