Fix typos which lead to a multiple of layout issues.
This fixes issues #50 and #52. The underlying cause was that
I renamed the 'layout' attribute to 'layout_manager' and didn't
update the overridden class traits on ScrollArea, Splitter, and
Tabbed to use the new name. I also neglected to update the
traverse_descendents method on ConstraintsLayout to use the
new name.