track the source text for certain constructs for documentation purposes.
This adds a __base__ attribute to EnamlDeclaration which is the python
source text which determines from what component a declaration is derived.
This allows auto-doc tools to create a nice representation of the derived
component.

This required updating the parser to track the source text for Python
ast nodes. The source is (at the moment) only tracked where it is useful
or otherwise trivial to compute. In other cases, it is simply set to
'<untracked>' instead of slowing down the parser to compute something
that will likely never be needed.