refactor the ConstraintsLayout manager and friends
The layout manager is now very simple. The constraints computation is now
split between the Container and LayoutComponent, with each computing the
constraints over which they have control. The sharing of constraints
between parent/child containers now follows a specific api, rather than
the smash-and-grab approach we were using before.