cache the size hint of the container since it is relatively expensive to compute
Since we advertise that a container can compute a size hint base on the layout manager and
independent of the toolkit widget, the toolkit widget is free to call into the function to
compute its own size hint. The widget may thus call this function *a lot*, like on every
resize even a lot, and so it must be cached in order to be fast enough. The cache is cleared
on every size hint updated event.