Remove the thread safety in the signaling module.
The justification here is that emitting signals from multiple threads
will be a rare case, and the common case shouldn't have to suffer
the performance penalty. For the case where thread safety is needed,
we can either add a thread safe signal in the future, or the user
can lock around their signaling operations.