Enhance the grammar to allow saving more vertical space.
This change allows the developer some more flexibility in conserving
vertical space for one-line definitions. The RHS of a declaration
can now have a one-liner attribute binding similar to how the RHS of
a :: operator can be a one-liner statement:

Label: text = 'foo'

CustomWidget: pass

This update also allows an in-line definition of an identifier. I
was getting many complaints that the `id:` tag was confusing. This
new form tries to be more indicative of what is going on:

Label: label1:
   text = 'foo'

Label: label1: text = 'foo'

CustomWidget: custom: pass

The old form of the `id:` tag is still supported. But mixing the two
on the same widget (declaring multiple ids) is a syntax error.

Users should be encouraged to adopt the new style.