Remove the validate_children functionality.
This was a mistake. It will become very difficult in the future to add
custom children that manipulate the tree in various ways. As an example,
imaging a MouseHandler component that can be used as the child of a
widget. It would be painful to go back and add that to the validator
for every object.